add getEditingDomain to handler so that it can be used for save
For save to work, the editing domain must be found at the handler level
rather than the factory. 
There are many different scenarios so we cannot cover all cases. The
default gets an editing domain from the source element if possible
otherwise creates a new one, but in some cases it might be better to
always create a new one.